requirements.txt
rembg==2.0.53
Pillow==10.1.0
tqdm==4.66.1
